Update issue 6489 Status: NeedsReview Be careful: branch is called issue-6489-2 Damn, I've been searching the content scripts several times for where it went wrong, but I was looking at the wrong point :(
Be a bit more restrictive in the part that handles the weird feed:https: protocol (and the //) (or actually, feed protocol with weird https: stuff and the //).
Please tell me if you know about URLs that match
normalProtocol:proto-col://*
as I am only aware of
feed:http(s)://
and therefore didn't allow an additional '-' in the 'second' protocol. I doubt they exist, but if they do, please tell me.
I like regex :D (except for in vb.NET due to the stupid implementation)